requirements.txt
# Core dependencies
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0

# HTTP & APIs
httpx>=0.27.0
slack-sdk>=3.27.0
openai>=1.12.0

# Utilities
pytz>=2024.1
python-dateutil>=2.8.2
rapidfuzz>=3.6.0
pyyaml>=6.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code Quality
ruff>=0.2.0
mypy>=1.8.0
black>=24.1.0

# Streamlit & Visualization
streamlit>=1.29.0
plotly>=5.17.0
